+++ src/main.rs
@@ -1,6 +1,7 @@
 extern crate base64;
 extern crate bip39;
 extern crate crypto;
+extern crate hex;
 extern crate regex;
 
 #[macro_use]
@@ -12,9 +13,12 @@ use crypto::scrypt;
 use regex::Regex;
 use std::io;
 
+/// Defines command line arguments
 fn arg_matches<'a>() -> ArgMatches<'a> {
     App::new("scrypt-rs")
         .about("Read passphrase (first line from stdin), normalize it (drop extra whitespace) and pass it to scrypt")
+        .arg(Arg::with_name("short").short("S").long("short").takes_value(false)
+            .help("Return hex encoded scrypt derivated key"))
         .arg(Arg::with_name("salt").short("s").long("salt").takes_value(true).default_value("")
             .help("Set salt"))
         .arg(Arg::with_name("logN").short("L").long("logn").takes_value(true).default_value("19")
@@ -28,6 +32,7 @@ fn arg_matches<'a>() -> ArgMatches<'a> {
         .get_matches()
 }
 
+/// Reads stdin and normalizes passphrase
 fn subcommand_dispatch(app_m: ArgMatches) {
     let mut input = String::new();
     io::stdin()
@@ -35,64 +40,91 @@ fn subcommand_dispatch(app_m: ArgMatches) {
         .expect("Failed to read passphrase");
     let pass = normalize_passphrase(&input);
     let salt = app_m.value_of("salt").unwrap().to_string();
-    println!("Salt: \"{}\"", &salt);
-    println!("Normalized passphrase: \"{}\"", pass);
-    run_scrypt(&app_m, &pass, &salt);
+    let params = Params::from_matches(&app_m);
+    run_scrypt(params, &pass, &salt);
 }
 
+/// Program entrypoint
 fn main() {
     subcommand_dispatch(arg_matches());
 }
 
+/// Parameters struct definition
 struct Params {
     log_n: u8,
     r: u32,
     p: u32,
     dk_len: usize,
+    short: bool,
 }
 
+/// Paramaters struct assignments from command line arguments values
 impl Params {
     fn from_matches(matches: &ArgMatches) -> Params {
         let log_n = value_t!(matches, "logN", u8).unwrap_or_else(|e| e.exit());
         let r = value_t!(matches, "r", u32).unwrap_or_else(|e| e.exit());
         let p = value_t!(matches, "p", u32).unwrap_or_else(|e| e.exit());
         let dk_len = value_t!(matches, "len", usize).unwrap_or_else(|e| e.exit());
+        let short = matches.is_present("short");
         Params {
             log_n,
             r,
             p,
             dk_len,
+            short,
         }
     }
 }
 
-fn derive_key(params: Params, pass: &str, salt: &str) -> Vec<u8> {
+/// Performs key derivation
+fn derive_key(params: &Params, pass: &str, salt: &str) -> Vec<u8> {
     let mut dk = vec![0; params.dk_len];
     let scrypt_params = scrypt::ScryptParams::new(params.log_n, params.r, params.p);
     scrypt::scrypt(pass.as_bytes(), salt.as_bytes(), &scrypt_params, &mut dk);
     dk
 }
 
-fn run_scrypt(app_m: &ArgMatches, pass: &str, salt: &str) {
-    let params = Params::from_matches(app_m);
-    let dk = derive_key(params, pass, salt);
-    print_hex("Scrypt: ", &dk);
-    let mnemonic = Mnemonic::from_entropy(&dk, Language::English).unwrap();
-    println!("BIP39: {}", mnemonic.phrase());
-    println!("base64: {}", base64::encode(&dk));
+/// Performs key derivation and outputs result
+fn run_scrypt(params: Params, pass: &str, salt: &str) {
+    let dk = derive_key(&params, pass, salt);
+    if params.short {
+        short_output(dk)
+    } else {
+        full_output(dk, pass, salt, params)
+    }
+}
+
+/// Minimal output returning the derived key in hexadecimal
+fn short_output(dk: Vec<u8>) {
+    println!("{}", hex::encode(dk));
+}
+
+/// Verbose output for derived key showing input parameters and generated key
+fn full_output(dk: Vec<u8>, pass: &str, salt: &str, params: Params) {
+    println!("Input | Salt: \"{}\"", &salt);
+    println!("Input | Normalized passphrase: \"{}\"", pass);
+    println!("Input | Scrypt parameters: cost factor {} - blocksize {} - parallelization {} - key length in bytes {}", params.log_n, params.r, params.p, params.dk_len);
+    println!(
+        "Output| Scrypt derived key in hexadecimal: {}",
+        hex::encode(&dk)
+    );
+    println!(
+        "Output| Scrypt derived key in base64: {}",
+        base64::encode(&dk)
+    );
+    match Mnemonic::from_entropy(&dk, Language::English) {
+        Ok(mnemonic) => println!(
+            "Output| Scrypt BIP39 words list representation: {}",
+            mnemonic.phrase()
+        ),
+        Err(_) => println!("Output| Scrypt BIP39: Unable to generate words list"),
+    };
 }
 
 // Utils
 
+/// Normalizes passphrase removing leading and trailing spaces
 fn normalize_passphrase(input: &str) -> String {
     let re = Regex::new(r"\s+").unwrap();
     re.replace_all(input, " ").trim().to_owned()
-}
-
-fn print_hex(prefix: &str, bytes: &[u8]) {
-    print!("{}", prefix);
-    for b in bytes.iter() {
-        print!("{:x}", b);
-    }
-    println!("\n");
 }