+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2021-03-13
+
+### Added
+
+- `-S` parameter to output only the derived key in hexadecimal.
+
+### Changed
+
+- Output refactored to differentiate more clearly what information are inputs and which ones are outputs.
+
 ## [0.1.0] - 2021-03-08
 
 ### Added